+++ Packages/com.unity.inputsystem/InputSystem/Actions/InputActionState.cs
@@ -1242,14 +1242,28 @@ namespace UnityEngine.InputSystem
                 memory.controlMagnitudes[triggerControlIndex] = trigger.magnitude;
             }
 
+            // Determine which control to consider the one currently associated with the action.
+            // We do the same thing as for the triggered control and in the case of a composite,
+            // switch to the first control of the composite.
+            var actionStateControlIndex = actionState->controlIndex;
+            if (bindingStates[actionState->bindingIndex].isPartOfComposite)
+            {
+                var compositeBindingIndex = bindingStates[actionState->bindingIndex].compositeOrCompositeBindingIndex;
+                actionStateControlIndex = bindingStates[compositeBindingIndex].controlStartIndex;
+            }
+
             // Never ignore state changes for actions that aren't currently driven by
             // anything.
-            if (actionState->controlIndex == kInvalidIndex)
+            if (actionStateControlIndex == kInvalidIndex)
             {
                 actionState->magnitude = trigger.magnitude;
                 Profiler.EndSample();
                 return false;
             }
+
+            // Find out if we get triggered from the control that is actively driving the action.
+            var isControlCurrentlyDrivingTheAction = triggerControlIndex == actionStateControlIndex ||
+                controls[triggerControlIndex] == controls[actionStateControlIndex];                                      // Same control, different binding.
 
             // If the control is actuated *more* than the current level of actuation we recorded for the
             // action, we process the state change normally. If this isn't the control that is already
@@ -1267,20 +1281,13 @@ namespace UnityEngine.InputSystem
                 // account or not.
                 // NOTE: For composites, we have forced triggerControlIndex to the first control
                 //       in the composite. See above.
-                if (trigger.magnitude > 0 && triggerControlIndex != actionState->controlIndex && actionState->magnitude > 0)
+                if (trigger.magnitude > 0 && !isControlCurrentlyDrivingTheAction && actionState->magnitude > 0)
                     actionState->hasMultipleConcurrentActuations = true;
 
                 // Keep recorded magnitude in action state up to date.
                 actionState->magnitude = trigger.magnitude;
                 Profiler.EndSample();
                 return false;
-            }
-
-            var actionStateControlIndex = actionState->controlIndex;
-            if (bindingStates[actionState->bindingIndex].isPartOfComposite)
-            {
-                var compositeBindingIndex = bindingStates[actionState->bindingIndex].compositeOrCompositeBindingIndex;
-                actionStateControlIndex = bindingStates[compositeBindingIndex].controlStartIndex;
             }
 
             // If the control is actuated *less* then the current level of actuation we
@@ -1291,14 +1298,15 @@ namespace UnityEngine.InputSystem
             {
                 // If we're not currently driving the action, it's simple. Doesn't matter that we lowered
                 // actuation as we didn't have the highest actuation anyway.
-                if (triggerControlIndex != actionStateControlIndex)
+                if (!isControlCurrentlyDrivingTheAction)
                 {
                     Profiler.EndSample();
                     ////REVIEW: should we *count* actuations instead? (problem is that then we have to reliably determine when a control
                     ////        first actuates; the current solution will occasionally run conflict resolution when it doesn't have to
                     ////        but won't require the extra bookkeeping)
                     // Do NOT let this control state change affect the action.
-                    actionState->hasMultipleConcurrentActuations = true;
+                    if (trigger.magnitude > 0)
+                        actionState->hasMultipleConcurrentActuations = true;
                     return true;
                 }
 
@@ -1435,20 +1443,13 @@ namespace UnityEngine.InputSystem
 
             // If we're not really effecting any change on the action, ignore the control state change.
             // NOTE: We may be looking at a control here that points in a completely direction, for example, even
-            //       though it has the same magnitude. However, we require a control to *higher* absolute actuation
+            //       though it has the same magnitude. However, we require a control to *increase* absolute actuation
             //       before we let it drive the action.
-            if (Mathf.Approximately(trigger.magnitude, actionState->magnitude))
+            if (!isControlCurrentlyDrivingTheAction && Mathf.Approximately(trigger.magnitude, actionState->magnitude))
             {
-                // However, if we have changed the control to a different control on the same composite, we *should* let
-                // it drive the action - this is like a direction change on the same control.
-                if (bindingStates[trigger.bindingIndex].isPartOfComposite && triggerControlIndex == actionStateControlIndex)
-                    return false;
                 // If we do have an actuation on a control that isn't currently driving the action, flag the action has
                 // having multiple concurrent inputs ATM.
-                // NOTE: We explicitly check for whether it is in fact not the same control even if the control indices are different.
-                //       The reason is that we allow the same control, on the same action to be bound more than once on the same
-                //       action.
-                if (trigger.magnitude > 0 && triggerControlIndex != actionState->controlIndex && controls[triggerControlIndex] != controls[actionState->controlIndex])
+                if (trigger.magnitude > 0)
                     actionState->hasMultipleConcurrentActuations = true;
                 return true;
             }
@@ -2284,6 +2285,8 @@ namespace UnityEngine.InputSystem
             Debug.Assert(bindingIndex >= 0 && bindingIndex < totalBindingCount, "Binding index is out of range");
             Debug.Assert(controlIndex >= 0 && controlIndex < totalControlCount, "Control index is out of range");
 
+            // If the control is part of a composite, it's the InputBindingComposite
+            // object that computes a magnitude for the whole composite.
             if (bindingStates[bindingIndex].isPartOfComposite)
             {
                 var compositeBindingIndex = bindingStates[bindingIndex].compositeOrCompositeBindingIndex;